part 1 initial split
